+++ py/moq-rs/moq/__init__.py
@@ -7,7 +7,15 @@ from moq_ffi import MoqError as Error
 
 from .client import Client, connect
 from .log import log_level
-from .origin import Announced, AnnouncedBroadcast, Announcement, OriginConsumer, OriginProducer
+from .origin import (
+    Announced,
+    AnnouncedBroadcast,
+    Announcement,
+    BroadcastRequest,
+    OriginConsumer,
+    OriginDynamic,
+    OriginProducer,
+)
 from .publish import (
     AudioProducer,
     BroadcastDynamic,
@@ -67,6 +75,7 @@ __all__ = [
     "BroadcastConsumer",
     "BroadcastDynamic",
     "BroadcastProducer",
+    "BroadcastRequest",
     "Catalog",
     "CatalogConsumer",
     "Client",
@@ -84,6 +93,7 @@ __all__ = [
     "MediaProducer",
     "MediaStreamProducer",
     "OriginConsumer",
+    "OriginDynamic",
     "OriginProducer",
     "Request",
     "Server",